# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: ai
    description: Gemini-powered cooking assistance
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /ai/fridge:
    post:
      operationId: analyzeFridgeImage
      tags: [ai]
      summary: Identify ingredients from a fridge image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FridgeAnalysisRequest"
      responses:
        "200":
          description: Ingredients and three recipe ideas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FridgeAnalysisResponse"
        "400":
          description: Invalid input
        "502":
          description: AI provider error
  /ai/recipe:
    post:
      operationId: lookupRecipe
      tags: [ai]
      summary: Look up a lighter recipe and nutrition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeLookupRequest"
      responses:
        "200":
          description: Recipe instructions and nutrition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeLookupResponse"
        "400":
          description: Invalid input
        "502":
          description: AI provider error
  /ai/meal-tray:
    post:
      operationId: generateMealTray
      tags: [ai]
      summary: Generate a full 3-dish mâm cơm (Đạm - Rau - Canh) tray for the family
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealTrayRequest"
      responses:
        "200":
          description: Mâm cơm 3 món with cost, hand-rule portions and health notes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealTrayResponse"
        "400":
          description: Invalid input
        "502":
          description: AI provider error
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SafetyContext:
      type: object
      properties:
        kids:
          type: integer
          minimum: 0
        allergies:
          type: array
          items:
            type: string
        allergyOther:
          type: string
      required:
        - kids
        - allergies
        - allergyOther
    FridgeAnalysisRequest:
      type: object
      properties:
        imageData:
          type: string
          description: Base64 data URL for an uploaded food or fridge image
        mimeType:
          type: string
        safety:
          $ref: "#/components/schemas/SafetyContext"
      required:
        - imageData
        - mimeType
        - safety
    FridgeIngredient:
      type: object
      properties:
        name:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
      required:
        - name
        - confidence
    NutritionSummary:
      type: object
      properties:
        calories:
          type: number
          minimum: 0
        protein:
          type: number
          minimum: 0
        fat:
          type: number
          minimum: 0
      required:
        - calories
        - protein
        - fat
    FridgeDish:
      type: object
      properties:
        name:
          type: string
        why:
          type: string
        ingredients:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: string
        nutrition:
          $ref: "#/components/schemas/NutritionSummary"
      required:
        - name
        - why
        - ingredients
        - steps
        - nutrition
    FridgeAnalysisResponse:
      type: object
      properties:
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/FridgeIngredient"
        dishes:
          type: array
          minItems: 3
          maxItems: 3
          items:
            $ref: "#/components/schemas/FridgeDish"
        safetyNotes:
          type: array
          items:
            type: string
      required:
        - ingredients
        - dishes
        - safetyNotes
    RecipeLookupRequest:
      type: object
      properties:
        dishName:
          type: string
          minLength: 1
        safety:
          $ref: "#/components/schemas/SafetyContext"
      required:
        - dishName
        - safety
    RecipeIngredient:
      type: object
      properties:
        name:
          type: string
        amount:
          type: string
      required:
        - name
        - amount
    RecipeLookupResponse:
      type: object
      properties:
        dishName:
          type: string
        servings:
          type: integer
          minimum: 1
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        steps:
          type: array
          items:
            type: string
        nutrition:
          $ref: "#/components/schemas/NutritionSummary"
        safetyNotes:
          type: array
          items:
            type: string
      required:
        - dishName
        - servings
        - ingredients
        - steps
        - nutrition
        - safetyNotes
    FamilyProfile:
      type: object
      properties:
        adults:
          type: integer
          minimum: 0
        elderly:
          type: integer
          minimum: 0
        kids:
          type: integer
          minimum: 0
        budgetPerMeal:
          type: integer
          minimum: 0
          description: Ngân sách ước tính cho mâm cơm này, tính bằng VNĐ (tùy chọn)
      required:
        - adults
        - elderly
        - kids
    BagIngredientInput:
      type: object
      description: Một nguyên liệu người dùng tự nhập ở chế độ Tự Nhập Túi Đồ / Dọn Tủ
      properties:
        name:
          type: string
        amount:
          type: string
      required:
        - name
        - amount
    MealTrayRequest:
      type: object
      properties:
        familyProfile:
          $ref: "#/components/schemas/FamilyProfile"
        safety:
          $ref: "#/components/schemas/SafetyContext"
        bagIngredients:
          type: array
          items:
            $ref: "#/components/schemas/BagIngredientInput"
        dateISO:
          type: string
          description: Ngày dương lịch hiện tại (YYYY-MM-DD) để tự nhận biết Mùng 1/Rằm âm lịch; mặc định lấy giờ máy chủ nếu bỏ trống
      required:
        - familyProfile
        - safety
    MealTrayIngredient:
      type: object
      properties:
        item:
          type: string
        amount:
          type: string
        cost:
          type: integer
          minimum: 0
      required:
        - item
        - amount
        - cost
    MealTrayDish:
      type: object
      properties:
        category:
          type: string
          enum: ["Món Đạm", "Món Rau", "Món Canh"]
        name:
          type: string
        portion_hand_rule:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/MealTrayIngredient"
      required:
        - category
        - name
        - portion_hand_rule
        - ingredients
    MealTrayResponse:
      type: object
      properties:
        meal_title:
          type: string
        total_estimated_cost:
          type: integer
          minimum: 0
        cooking_time_minutes:
          type: integer
          minimum: 0
        health_benefits_note:
          type: string
        dishes:
          type: array
          minItems: 3
          maxItems: 3
          items:
            $ref: "#/components/schemas/MealTrayDish"
        tags:
          type: array
          items:
            type: string
      required:
        - meal_title
        - total_estimated_cost
        - cooking_time_minutes
        - health_benefits_note
        - dishes
        - tags