# lib/api-zod/src/generated/types/bagIngredientInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Một nguyên liệu người dùng tự nhập ở chế độ Tự Nhập Túi Đồ / Dọn Tủ
 */
export interface BagIngredientInput {
  name: string;
  amount: string;
}