# lib/api-zod/src/generated/types/familyProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FamilyProfile {
  /** @minimum 0 */
  adults: number;
  /** @minimum 0 */
  elderly: number;
  /** @minimum 0 */
  kids: number;
  /**
     * Ngân sách ước tính cho mâm cơm này, tính bằng VNĐ (tùy chọn)
     * @minimum 0
     */
  budgetPerMeal?: number;
}