# lib/api-zod/src/generated/types/mealTrayDish.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MealTrayDishCategory } from './mealTrayDishCategory';
import type { MealTrayIngredient } from './mealTrayIngredient';

export interface MealTrayDish {
  category: MealTrayDishCategory;
  name: string;
  portion_hand_rule: string;
  ingredients: MealTrayIngredient[];
}