# lib/api-zod/src/generated/types/mealTrayIngredient.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MealTrayIngredient {
  item: string;
  amount: string;
  /** @minimum 0 */
  cost: number;
}