# lib/api-zod/src/generated/types/mealTrayRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BagIngredientInput } from './bagIngredientInput';
import type { FamilyProfile } from './familyProfile';
import type { SafetyContext } from './safetyContext';

export interface MealTrayRequest {
  familyProfile: FamilyProfile;
  safety: SafetyContext;
  bagIngredients?: BagIngredientInput[];
  /** Ngày dương lịch hiện tại (YYYY-MM-DD) để tự nhận biết Mùng 1/Rằm âm lịch; mặc định lấy giờ máy chủ nếu bỏ trống */
  dateISO?: string;
}