# lib/api-zod/src/generated/types/mealTrayResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MealTrayDish } from './mealTrayDish';

export interface MealTrayResponse {
  meal_title: string;
  /** @minimum 0 */
  total_estimated_cost: number;
  /** @minimum 0 */
  cooking_time_minutes: number;
  health_benefits_note: string;
  /**
     * @minItems 3
     * @maxItems 3
     */
  dishes: MealTrayDish[];
  tags: string[];
}